Added new errors to all commands

+++ src/commands/taskkill.js
@@ -1,32 +1,35 @@
 const chalk = require("chalk");
 
+const _errorInterpret = require("../functions/errorInt");
+
 const taskkill = (pid) => {
   if (typeof pid === "undefined") {
-    console.log(`Please enter a PID to kill. Example: ${chalk.yellow("taskkill 1234")}.`);
-    console.log();
-
+    _errorInterpret("0x0041");
     return;
   }
 
-  console.log(`Killing PID ${chalk.yellow(pid)}...`);
+  const isNumeric = (str) => {
+    if (typeof str !== "string") return false;
+    return !isNaN(str) && !isNaN(parseFloat(str));
+  };
+
+  if (!isNumeric(pid)) {
+    _errorInterpret("0x0042");
+    return;
+  }
+
+  console.log(`Killing process with PID ${chalk.yellow(pid)}...`);
 
   try {
     process.kill(Number(pid));
+    console.log(`Successfully killed process ${chalk.green(pid)}.\n`);
   } catch (err) {
-    if (err.message.toLowerCase().includes("eperm")) {
-      console.log(`You do not have permission to kill ${chalk.red(`${pid}`)}.`);
-      console.log();
+    if (err.code === "EPERM") {
+      _errorInterpret("0x0043", { variable: pid });
     } else {
-      console.log(`${chalk.red(err.message)}.`);
-      console.log();
+      _errorInterpret("0x0044", { variable: pid, wordCode: err.code });
     }
-
-    return;
   }
-
-  console.log(`Successfully killed process ${chalk.green(pid)}.`);
-
-  console.log();
 };
 
 module.exports = taskkill;
